+++ git-branchless/tests/command/test_query.rs
@@ -84,7 +84,7 @@ fn test_query_eval_error() -> eyre::Result<()> {
             },
         )?;
         insta::assert_snapshot!(stderr, @r###"
-        Evaluation error for expression 'foo()': no function with the name 'foo' could be found; these functions are available: all, ancestors, ancestors.nth, author.date, author.email, author.name, branches, children, committer.date, committer.email, committer.name, descendants, difference, draft, heads, intersection, message, none, not, only, parents, parents.nth, paths.changed, range, roots, stack, union
+        Evaluation error for expression 'foo()': no function with the name 'foo' could be found; these functions are available: all, ancestors, ancestors.nth, author.date, author.email, author.name, branches, children, committer.date, committer.email, committer.name, descendants, difference, draft, exactly, heads, intersection, message, none, not, only, parents, parents.nth, paths.changed, range, roots, stack, union
         "###);
         insta::assert_snapshot!(stdout, @"");
     }